updated tb_gfn with new API
