+++ chunkgfn/algo/tb_gfn_.py
@@ -1,0 +1,60 @@
+import torch
+import torch.nn.functional as F
+from torch.distributions import Categorical
+
+from chunkgfn.gfn.base_unconditional_gfn import UnConditionalSequenceGFN
+
+from ..constants import NEGATIVE_INFINITY
+
+
+class TBGFN(UnConditionalSequenceGFN):
+    def compute_loss(self, trajectories, actions, dones, logreward):
+        """Compute the loss for the model.
+        Args:
+            trajectories (torch.Tensor[batch_size, trajectory_length, *state_shape]): Trajectories for each sample in the batch.
+            actions (torch.Tensor[batch_size, trajectory_length]): Actions for each sample in the batch.
+            dones (torch.Tensor[batch_size, trajectory_length]): Whether the trajectory is done or not.
+            logreward (torch.Tensor[batch_size]): Log reward.
+        Return:
+            loss (torch.Tensor[1]): Loss.
+            logZ (torch.Tensor[1]): Log partition function.
+        """
+        log_pf = 0
+        log_pb = 0
+        for t in range(trajectories.shape[1]):
+            state = trajectories[:, t]
+            logit_pf = self.get_forward_logits(state)
+            forward_mask = self.trainer.datamodule.get_forward_mask(state)
+            logit_pf = torch.where(
+                forward_mask,
+                logit_pf,
+                torch.tensor(NEGATIVE_INFINITY).to(logit_pf),
+            )
+            if t < trajectories.shape[1] - 1:
+                log_pf += (Categorical(logits=logit_pf).log_prob(actions[:, t])) * (
+                    ~dones[:, t] + 0
+                )
+            if t > 0:
+                backward_actions = self.trainer.datamodule.get_parent_actions(state)
+                logp_b_s = torch.where(
+                    backward_actions == 1, torch.tensor(0.0), -torch.inf
+                ).to(logit_pf)
+                # When no action is available, just fill with uniform because it won't be picked anyway in the backward_step. Doing this avoids having nan when computing probabilities
+                logp_b_s = torch.where(
+                    (logp_b_s == -torch.inf).all(dim=-1).unsqueeze(1),
+                    torch.tensor(0.0),
+                    logp_b_s,
+                )
+                log_pb += torch.where(
+                    dones[:, t] | self.trainer.datamodule.is_initial_state(state),
+                    torch.tensor(0.0),
+                    Categorical(logits=logp_b_s).log_prob(actions[:, t - 1]),
+                )
+
+        logZ = self.logZ
+
+        loss = F.mse_loss(
+            logZ + log_pf, (logreward / self.hparams.reward_temperature) + log_pb
+        )
+
+        return loss, logZ
